Unable to stub so moved a helper where it belongs and can now test in peace
