+++ src/modules/mhct-lookup.js
@@ -2,6 +2,7 @@ const fetch = require('node-fetch');
 const Logger = require('../modules/logger');
 const { DateTime, Duration } = require('luxon');
 const { calculateRate, prettyPrintArrayAsString, intToHuman } = require('../modules/format-utils');
+const { getSearchedEntity } = require('../modules/search-helpers');
 const { MessageEmbed } = require('discord.js');
 
 const refresh_rate = Duration.fromObject({ minutes: 5 });
@@ -211,30 +212,6 @@ async function formatMice(isDM, mouse, opts) {
     reply += prettyPrintArrayAsString(attracts, columnFormatting, headers, '=');
     reply += '```\n' + `HTML version at: ${target_url}`;
     return reply;
-}
-
-/**
- * Return a sorted list of approximate matches to the given input and container
- *
- * @param {string} input The text to match against
- * @param {Array} values An array of objects with a lowerValue property.
- * @returns {Array <number>[]} Up to 10 indices and their search score.
- */
-function getSearchedEntity(input, values) {
-    if (!input.length || !Array.isArray(values) || !values || !values.length)
-        return [];
-
-    const matches = values.filter(v => v.lowerValue.includes(input)).map(v => {
-        return { entity: v, score: v.lowerValue.indexOf(input) };
-    });
-    matches.sort((a, b) => {
-        const r = a.score - b.score;
-        // Sort lexicographically if the scores are equal.
-        return r ? r : a.entity.value.localeCompare(b.entity.value, { sensitivity: 'base' });
-    });
-    // Keep only the top 10 results.
-    matches.splice(10);
-    return matches.map(m => m.entity);
 }
 
 /**
