+++ tests/modules/mhct-lookup.js
@@ -1,6 +1,8 @@
 const test = require('tape');
 const sinon = require('sinon');
 
+const searchHelper = require('../../src/modules/search-helpers');
+const getSearchedEntityStub = sinon.stub(searchHelper, 'getSearchedEntity').returns([]);
 const mhct_lookup = require('../../src/modules/mhct-lookup');
 
 //const getMHCTList = mhct_lookup.getMHCTList;
@@ -28,7 +30,25 @@ test('getFilter', suite => {
         sinon.reset();
 
     });
-    suite.test('given input that can\'t be turned into a truthy string - returns undefined', t => {
+    suite.test('given string input - returns known shortcuts', t => {
+        const inputs = [
+            { input: '3', expected: '3_days', },
+            { input: '3day', expected: '3_days', },
+            { input: 'all', expected: 'alltime', },
+            { input: 'allowance', expected: 'alltime', },
+            { input: 'current', expected: '1_month', }, //NOTE this can only be asserted because we don't load the filter list
+        ];
+        t.plan(inputs.length*2);
+        getSearchedEntityStub.returns([]);
+        inputs.forEach(input => {
+            const result = getFilter(input.input);
+            t.true(getSearchedEntityStub.calledOnce, 'should call search entity');
+            t.deepEqual(getSearchedEntityStub.args[0][0], input.expected, `should search for known shortcut ${input.input} = ${input.expected}`);
+            console.log(`Recived ${result}`);
+            getSearchedEntityStub.resetHistory();
+        });
+        sinon.reset();
+    });    suite.test('given input that can\'t be turned into a truthy string - returns undefined', t => {
         const inputs = [
             '',
             undefined,
